+++ assets/release_notes.md
@@ -4,6 +4,7 @@
 - Add ability to scan in received items using supplier barcodes
 - Store API token, rather than username:password
 - Ensure that user will lose access if token is revoked by server
+- Improve scroll-to-refresh behaviour across multiple widgets
 
 
 ### 0.12.8 - September 2023