+++ assets/release_notes.md
@@ -1,3 +1,9 @@
+### 0.14.0 - November 2023
+---
+
+- Adds option to pause and resume barcode scanning with camera
+- Adds option for "single shot" barcode scanning with camera
+
 ### 0.13.0 - October 2023
 ---